win32-test/requirements.txt
pywin32
pykiwoom
python-telegram-bot
python-dotenv
requests
beautifulsoup4
html5lib